update to babel 7: fix token conversion, fix breaking change with RestProperty/SpreadProperty
