+++ .babelrc
@@ -3,7 +3,6 @@
   "plugins": [
     "transform-class-properties",
     "transform-flow-strip-types",
-    "transform-object-rest-spread",
-    "transform-runtime"
+    "transform-object-rest-spread"
   ]
 }
