+++ .travis.yml
@@ -13,7 +13,6 @@ script:
 node_js:
   - "4"
   - "6"
-  - "7"
   - "8"
 
 matrix:
