+++ package.json
@@ -30,15 +30,14 @@
     "postpublish": "rm -Rf lib"
   },
   "devDependencies": {
-    "babel-cli": "^6.6.5",
-    "babel-helper-fixtures": "^6.18.2",
-    "babel-eslint": "~7.0.0",
-    "babel-plugin-transform-class-properties": "^6.6.0",
-    "babel-plugin-transform-flow-strip-types": "^6.7.0",
-    "babel-plugin-transform-object-rest-spread": "^6.6.5",
-    "babel-plugin-transform-runtime": "^6.9.0",
-    "babel-preset-es2015": "^6.6.0",
-    "babel-register": "^6.9.0",
+    "babel-cli": "7.0.0-alpha.18",
+    "babel-helper-fixtures": "7.0.0-alpha.18",
+    "babel-eslint": "8.0.0-alpha.17",
+    "babel-plugin-transform-class-properties": "7.0.0-alpha.18",
+    "babel-plugin-transform-flow-strip-types": "7.0.0-alpha.18",
+    "babel-plugin-transform-object-rest-spread": "7.0.0-alpha.18",
+    "babel-preset-es2015": "7.0.0-alpha.18",
+    "babel-register": "7.0.0-alpha.18",
     "benchmark": "^2.1.0",
     "chai": "^3.5.0",
     "chalk": "^1.1.1",
@@ -52,8 +51,7 @@
     "mocha": "^3.1.2"
   },
   "dependencies": {
-    "babel-runtime": "^6.9.2",
-    "babylon": "^6.8.1",
+    "babylon": "7.0.0-beta.19",
     "source-map-support": "^0.4.0"
   },
   "files": [
