+++ src/visitorKeys.js
@@ -69,11 +69,9 @@ export default {
     Program: ['body', 'directives'],
     RegExpLiteral: [],
     RestElement: ['argument'],
-    RestProperty: ['argument'],
     ReturnStatement: ['argument'],
     SequenceExpression: ['expressions'],
     SpreadElement: ['argument'],
-    SpreadProperty: ['argument'],
     StringLiteral: [],
     Super: [],
     SwitchCase: ['test', 'consequent'],
